auth done, tested. all configurations are done. time to the core now.

+++ mcp-server/src/main/java/com/g4/chatbot/repos/VerificationTokenRepository.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface VerificationTokenRepository extends JpaRepository<VerificationToken, Integer> {
+public interface VerificationTokenRepository extends JpaRepository<VerificationToken, Long> {
     Optional<VerificationToken> findByToken(String token);
-    Optional<VerificationToken> findByUserIdAndUserType(Integer userId, String userType);
+    Optional<VerificationToken> findByUserIdAndUserType(Long userId, String userType);
     Optional<VerificationToken> findFirstByUserTypeOrderByCreatedDateDesc(String userType);
 }
